+++ app/src/main/java/se/juneday/systemetappbasic/MainActivity.java
@@ -5,25 +5,24 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
-import android.widget.RatingBar;
 import android.widget.Toast;
 
 public class MainActivity extends AppCompatActivity {
     private Button Button2;
     private Button button1;
-    private Button buttonsubmit;
-    private RatingBar rating_b;
+
+
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        onButtonClickListener();
+
 
 
         Button2 = (Button) findViewById(R.id.button2);
         button1 = (Button) findViewById(R.id.button1);
-        rating_b =(RatingBar) findViewById(R.id.ratingbar);
+
 
     }
 
@@ -56,17 +55,5 @@ public class MainActivity extends AppCompatActivity {
         });
     }
 
-    public void onButtonClickListener() {
-        rating_b = (RatingBar) findViewById(R.id.ratingbar);
-        buttonsubmit = (Button) findViewById(R.id.button3);
 
-        buttonsubmit.setOnClickListener(
-                new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        Toast.makeText(MainActivity.this, getString(R.string.thanksToast), Toast.LENGTH_SHORT).show();
-                    }
-                }
-        );
-    }
 }
